TMP Fix for getRunner
There is a bug with the GitHub API:
/repos/{owner}/{repo}/actions/runners does not return headers.link,
which causes pagination to fail and only returns the first 30 items
(the default per_page value).

Temporarily, the per_page limit has been increased to the maximum
allowed value (100) to try and avoid cases where the runner is not
returned in the first batch.

This fix will be in place until the bug is resolved on GitHub's side:
https://github.com/orgs/community/discussions/144871